Can't get GitHub work
We decided to use a USB stick to share and merge our works, as GitHub is
not compatible with Greenfoot and it overwrites files, syncs wrong
content and uploads temporary files that can crash other member's
Greenfoot. I will still try to push bigger changes here, so it can be
trackable.